typefindhelper: Fix signed integer overflow
Make sure the whole calculation is done with 64bit unsigned values

(To be ready for people want to typefind exabyte files).

+++ libs/gst/base/gsttypefindhelper.c
@@ -446,7 +446,7 @@ buf_helper_find_peek (gpointer data, gint64 off, guint size)
     return NULL;
   }
 
-  if ((off + size) <= helper->size)
+  if (((guint64) off + size) <= helper->size)
     return helper->data + off;
 
   return NULL;
